requirements.txt
# Testing Framework
pytest==7.4.3
pytest-html==4.1.1
pytest-xdist==3.3.1

# Web Automation
playwright==1.40.0

# Reporting and Logging
allure-pytest==2.13.2
loguru==0.7.2

# Configuration Management
pydantic==2.5.0
pyyaml==6.0.1

# Test Data
faker==20.1.0

# Code Quality
black==23.11.0
flake8==6.1.0 